separated MainWindow.py file into two file EpicSevenAutomationMain and DeviceSelectionUI placed under ui pacakge, and update Main.py accordingly
